refactor: remove dead code (#4016)
