+++ engine/src/main/java/org/terasology/math/Direction.java
@@ -17,7 +17,6 @@
 package org.terasology.math;
 
 import com.google.common.collect.Maps;
-import org.joml.Vector3fc;
 import org.terasology.math.geom.Vector3f;
 import org.terasology.math.geom.Vector3i;
