+++ engine/src/main/java/org/terasology/math/MatrixUtils.java
@@ -21,7 +21,6 @@ import org.joml.Matrix4fc;
 import org.lwjgl.BufferUtils;
 import org.terasology.math.geom.Matrix3f;
 import org.terasology.math.geom.Matrix4f;
-import org.terasology.math.geom.Vector3f;
 
 import java.nio.FloatBuffer;
