+++ engine/src/main/java/org/terasology/persistence/typeHandling/PersistedDataSerializer.java
@@ -21,7 +21,6 @@ import gnu.trove.iterator.TIntIterator;
 import gnu.trove.iterator.TLongIterator;
 
 import java.nio.ByteBuffer;
-import java.util.Collection;
 import java.util.Map;
 
 /**
