+++ engine/src/main/java/org/terasology/persistence/typeHandling/TypeHandlerContext.java
@@ -15,7 +15,6 @@
  */
 package org.terasology.persistence.typeHandling;
 
-import org.reflections.Reflections;
 import org.terasology.persistence.typeHandling.reflection.SerializationSandbox;
 import org.terasology.reflection.TypeInfo;
