+++ engine/src/main/java/org/terasology/persistence/typeHandling/TypeHandlerFactory.java
@@ -18,7 +18,6 @@ package org.terasology.persistence.typeHandling;
 import org.terasology.reflection.TypeInfo;
 
 import java.util.Optional;
-import java.util.function.BiFunction;
 
 /**
  * Creates type handlers for a set of types. Type handler factories are generally used when a set of types
