+++ engine/src/main/java/org/terasology/persistence/typeHandling/mathTypes/Rect2fTypeHandler.java
@@ -19,7 +19,6 @@ package org.terasology.persistence.typeHandling.mathTypes;
 import com.google.common.collect.Maps;
 import org.terasology.math.geom.Rect2f;
 import org.terasology.math.geom.Vector2f;
-import org.terasology.persistence.typeHandling.DeserializationException;
 import org.terasology.persistence.typeHandling.PersistedData;
 import org.terasology.persistence.typeHandling.PersistedDataMap;
 import org.terasology.persistence.typeHandling.PersistedDataSerializer;
