+++ engine/src/main/java/org/terasology/persistence/typeHandling/mathTypes/Rect2iTypeHandler.java
@@ -20,7 +20,6 @@ import com.google.common.collect.Maps;
 
 import org.terasology.math.geom.Rect2i;
 import org.terasology.math.geom.Vector2i;
-import org.terasology.persistence.typeHandling.DeserializationException;
 import org.terasology.persistence.typeHandling.PersistedData;
 import org.terasology.persistence.typeHandling.PersistedDataMap;
 import org.terasology.persistence.typeHandling.PersistedDataSerializer;
