+++ engine/src/main/java/org/terasology/physics/engine/PhysicsEngine.java
@@ -20,7 +20,6 @@ import org.terasology.entitySystem.entity.EntityRef;
 import org.terasology.math.geom.Vector3f;
 import org.terasology.physics.CollisionGroup;
 import org.terasology.physics.Physics;
-import org.terasology.physics.shapes.CollisionShapeFactory;
 
 import java.util.Iterator;
 import java.util.List;
