+++ engine/src/main/java/org/terasology/rendering/nui/widgets/UIRadialRing.java
@@ -15,8 +15,6 @@
  */
 package org.terasology.rendering.nui.widgets;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.terasology.math.geom.Rect2i;
 import org.terasology.math.geom.Vector2i;
 import org.terasology.rendering.nui.BaseInteractionListener;
